% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actividad_economica_de_la_madre}
\alias{actividad_economica_de_la_madre}
\title{Nacimientos por actividad económica de la madre}
\format{
30 renglones x 4 columnas

\describe{
\item{Año}{Año del registro}
\item{Trabajo_remunerado}{Porcentaje de nacimientos cuyo progenitora tiene trabajo remunerado}
\item{Trabajo_no_remunerado}{Porcentaje de nacimientos cuyo progenitora tiene trabajo no remunerado}
\item{Trabajo_no_especificado}{Porcentaje de nacimientos cuyo progenitora no especifíca su fuente de trabajo}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
actividad_economica_de_la_madre
}
\description{
Descripción detallada del dataset
}
\keyword{datasets}
