% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actividad_economica_del_progenitor}
\alias{actividad_economica_del_progenitor}
\title{Actividad económica de la padre}
\format{
30 renglones x 4 columnas

\describe{
\item{Año}{Año del registro}
.\item{Trabajo_remunerado}{Porcentaje de nacimientos cuyo progenitor tiene trabajo remunerado}
\item{Trabajo_no_remunerado}{Porcentaje de nacimientos cuyo progenitor tiene trabajo remunerado}
\item{Trabajo_no_especificado}{Porcentaje de nacimientos cuyo progenitor no especifica su fuente de trabajo}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
actividad_economica_del_progenitor
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por condición de actividad económica, 1990-2019
}
\keyword{datasets}
