% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curso.feminismo.datos-package.R
\docType{package}
\name{curso.feminismo.datos-package}
\alias{curso.feminismo.datos}
\alias{curso.feminismo.datos-package}
\title{curso.feminismo.datos: Conjunto de datos para el curso de Feminismo de Datos}
\description{
Este paquete recoge algunos conjuntos de datos abiertos que proporciona el gobierno mexicano que se usaran 
  en el curso iintersemestral 2020-02 de Feminismo de datos en la facultad de Psicologia.
  Contiene set de datos para fecundidad en niñas y adolescentes, brecha de genero e Interrupcion legal del embarazo.
  Cada uno de estos paquetes puede tiene su apartado de ayuda con la descripción de sus componentes.
}
\author{
\strong{Maintainer}: info AprendR \email{info@aprendr.org}

}
\keyword{internal}
