% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edad_progenitor}
\alias{edad_progenitor}
\title{Edad del progenitor}
\format{
30 renglones x 8 columnas

\describe{
\item{Año}{Año del registro}
\item{Menor_15}{hombres menores a 15 años}
\item{De_15-19}{hombres con edad en este rango}
\item{De_20-24}{hombres con edad en este rango}
\item{De_25_29}{hombres con edad en este rango}
\item{De_30_34}{hombres con edad en este rango}
\item{De_35_más}{hombres con edad en este rango}
\item{No_específicado}{No se registró el dato}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
edad_progenitor
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por edad del progenitor de sus hijos(as), 1990-2019
}
\keyword{datasets}
