% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{escolaridad_madre}
\alias{escolaridad_madre}
\title{Escolaridad de la madre}
\format{
30 renglones x 7 columnas

\describe{
\item{Año}{Año del registro}
\item{Sin_escolaridad}{Mujeres sin escolaridad}
\item{Primaria_incompleta}{Mujeres con primaria incomplera}
\item{Primaria_completa}{Mujeres con primaria}
\item{Secundaria_equivalente}{Mujeres con secundaria}
\item{Otra}{Otro tipo de educación}
\item{No_especificado}{No se registró no identificó}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
escolaridad_madre
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por escolaridad, 1990-2019
}
\keyword{datasets}
