% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{escolaridad_progenitor}
\alias{escolaridad_progenitor}
\title{Escolaridad de la padre}
\format{
30 renglones x 7 columnas

\describe{
\item{Año}{Año del registro}
\item{Sin_escolaridad}{Hombres sin escolaridad}
\item{Primaria_incompleta}{Hombres con primaria incomplera}
\item{Primaria_completa}{Hombres con primaria}
\item{Secundaria_equivalente}{Hombres con secundaria}
\item{Otra}{Otro tipo de educación}
\item{No_especificado}{No se registró no identificó}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
escolaridad_progenitor
}
\description{
República Mexicana. Distribución porcentual de nacimientos de niñas y adolescentes de 10 a 14 años de edad por escolaridad del progenitor de sus hijos(as), 1990-2019
}
\keyword{datasets}
