% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nacimiento_general}
\alias{nacimiento_general}
\title{Base de dados de fecundidad anual por entidad}
\format{
1023 renglones x 14 columnas

\describe{
\item{Año}{Año del registro}
\item{Clave Entidad}{Identificador de la entidad}
\item{Nacimientos 10-14}{Número de nacimientos entre 10 y 14 años}
\item{Nacimientos 12-14}{Número de nacimietos entre 12 y 14 años}
\item{Nacimientos 10}{Número de nacimientos a los 10 años}
\item{Nacimientos 11}{Número de nacimientos a los 11 años}
\item{Nacimientos 12}{Número de nacimientos a los 12 años}
\item{Nacimientos 13}{Número de nacimientos a los 13 años}
\item{Nacimientos 14}{Número de nacimientos a los 14 años}
\item{Adoloscentes_10-14}{Número de adolescentes entre 10 y 14 años}
\item{Adoloescentes_12-14}{Número de adolescentes entre 12 y 14 años}
\item{Razón_fecundidad_10-14}{Proporcion de nacimientos entre 10 y 14 años}
\item{Razón_fecundidad_12-14}{Proporcion de nacimientos entre 12 y 14 años}
}
}
\source{
\url{https://www.gob.mx/conapo/documentos/fecundidad-en-ninas-y-adolescentes-de-10-a-14-anos}
}
\usage{
nacimiento_general
}
\description{
Descripción detallada del dataset
}
\keyword{datasets}
