% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nacimiento_lugar_residencia}
\alias{nacimiento_lugar_residencia}
\title{Nacimientos por lugar de residencia}
\format{
30 renglones x 4 columnas

\describe{
\item{Año}{Año de registro}
\item{Área_rural}{Número de nacimientos por área rural}
\item{Área_urbana}{Número de nacimientos por área urbana}
\item{No_especificado}{Área no especificada}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
nacimiento_lugar_residencia
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por lugar de residencia, 1990-2019
}
\keyword{datasets}
