% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nacimientos_actividad_economica}
\alias{nacimientos_actividad_economica}
\title{Nacimientos por actividad económica}
\format{
30 renglones x 4 columnas

\describe{
\item{Año}{Año del registro}
\item{Trabajo_remunerado}{Porcentaje de nacimientos cuyo progenitor tiene trabajo remunerado}
\item{Trabajo_no_remunerado}{Porcentaje de nacimientos cuyo progenitor tiene trabajo no remunerado}
\item{No_especificado}{Porcentaje de nacimientos cuyo progenitor no especifíca su fuente de trabajo}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
nacimientos_actividad_economica
}
\description{
Descripción detallada del dataset
}
\keyword{datasets}
