% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numero_hijos}
\alias{numero_hijos}
\title{Número de hijos.}
\format{
30 renglones x 6 columnas

\describe{
\item{Año}{Año del registro}
\item{Con_1_hijo(a)}{Mujeres con un hija/o}
\item{Con_2_hijos(as)}{Mujeres con 2 hija/os}
\item{Con_3_hijos(as)}{Mujeres con 3 hija/os}
\item{Con_4_hijos(as)}{Mujeres con 4 hija/os}
\item{No_especificado}{No se registró o se identificó}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
numero_hijos
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por orden de nacimiento del último(a) hijo(a) nacido(a) vivo(a), 1990-2019
}
\keyword{datasets}
