% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registro_agente}
\alias{registro_agente}
\title{Titulo del dataset.}
\format{
Dimensiones del dataset

\describe{
\item{Nombre Columna 1}{definition}
\item{Nombre Columna 2}{definition}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
registro_agente
}
\description{
Descripción detallada del dataset
}
\keyword{datasets}
