% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registro_padres}
\alias{registro_padres}
\title{Declaración o registro de nacimiento}
\format{
11 renglones x 7 columnas

\describe{
\item{Año}{Año del registro}
\item{Ambos}{definition}
\item{Madre}{definition}
\item{Padre}{definition}
\item{Resgistrado}{definition}
\item{Otra_persona}{definition}
\item{No_especificado}{definition}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
registro_padres
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por persona que declaró el nacimiento, 1990-2019
}
\keyword{datasets}
