% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacion_conyugal}
\alias{situacion_conyugal}
\title{Situación conyugal}
\format{
30 renglones x 8 columnas

\describe{
\item{Año}{Año del registro}
\item{Soltera}{Porcentaje de nacimientos }
\item{Casada}{Porcentaje de nacimientos con situación conyugal casada}
\item{Unión_libre}{Porcentaje de nacimientos con situación conyugal casada}
\item{Separada}{Porcentaje de nacimientos con situación conyugal separada}
\item{Divorciada}{Porcentaje de nacimientos con situación conyugal divorciada}
\item{Viuda}{Porcentaje de nacimientos con situación conyugal viuda}
\item{No_especificado}{Porcentaje de nacimientos con situación conyugal sin especificar}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
situacion_conyugal
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por situación conyugal, 1990-2019
}
\keyword{datasets}
