% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacion_laboral_del_progenitor}
\alias{situacion_laboral_del_progenitor}
\title{Situación laboral del progenitor}
\format{
11 renglones x 6 columnas

\describe{
\item{Año}{Año del registro}
\item{Quehaceres_hogar}{Porcentaje del progenitor que realizan quehaceres del hogar}
\item{Estudiante}{Porcentaje del progenitor que estan realizando estudios}
\item{Incapacitado_para_trabajar}{Porcentaje del progenitor incapacitados para trabajar}
\item{Jubilado_pensionado}{Porcentaje del progenitor jubilados y/o pensionados}
\item{Otra}{Progenitor que realizan otras actividades}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
situacion_laboral_del_progenitor
}
\description{
República Mexicana. Distribución porcentual de nacimientos de niñas y adolescentes de 10 a 14 años de edad por tipo de ocupación del progenitor de sus hijos(as), 2009-2019
}
\keyword{datasets}
