% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situacion_laboral_madre}
\alias{situacion_laboral_madre}
\title{Situación laboral de la madre}
\format{
11 renglones x 5 columnas

\describe{
\item{Año}{Año del registro}
\item{Quehaceres_hogar}{Porcentaje de mujeres que realizan quehaceres del hogar}
\item{Estudiante}{Porcentaje de mujeres que estan realizando estudios}
\item{Incapacitado_para_trabajar}{Porcentaje de mujeres incapacitados para trabajar}
\item{Jubilado_pensionado}{Porcentaje del mujeres jubilados y/o pensionados}
\item{Otra}{Progenitor que realizan otras actividades}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
situacion_laboral_madre
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por tipo de ocupación, 2009-2019
}
\keyword{datasets}
