% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasa_fecundidad_residencia}
\alias{tasa_fecundidad_residencia}
\title{Tasa de fecunidad y residencia}
\format{
6 renglones x 5 columnas

\describe{
\item{Año}{Año del registro}
\item{Área_rural_10-14...2}{Tasa de fecunidad de mujeres de 10-14 años en área rural}
\item{Área_rural_12-14...3}{Tasa de fecunidad de mujeres de 12-14 añosen área rural}
\item{Área_rural_10-14...4}{Tasa de fecunidad de mujeres de 10-14 años en área urbana}
\item{Área_rural_12-14...5}{Tasa de fecunidad de mujeres de 12-14 añosen área urbana}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
tasa_fecundidad_residencia
}
\description{
República Mexicana. Tasa y razón de fecundidad en niñas y adolescentes de 10 a 14 años de edad por lugar de residencia, 2015-2020
}
\keyword{datasets}
