% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nacimiento_situacion_conyugal}
\alias{nacimiento_situacion_conyugal}
\title{Nacimiento por situación conyugal}
\format{
30 renglones x 8 columnas

\describe{
\item{Año}{Año del registro}
\item{Soltera}{Porcentaje de nacimientos }
\item{Casada}{Porcentaje de nacimientos con situación conyugal casada}
\item{Unión_libre}{Porcentaje de nacimientos con situación conyugal casada}
\item{Separada}{Porcentaje de nacimientos con situación conyugal separada}
\item{Divorciada}{Porcentaje de nacimientos con situación conyugal divorciada}
\item{Viuda}{Porcentaje de nacimientos con situación conyugal viuda}
\item{No_especificado}{Porcentaje de nacimientos con situación conyugal sin especificar}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
nacimiento_situacion_conyugal
}
\description{
Descripción detallada del dataset
}
\keyword{datasets}
