% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registro_agente}
\alias{registro_agente}
\title{Agente_que_atendió}
\format{
30 renglones x 5 columnas

\describe{
\item{Año}{Año del registro}
\item{Médico(a)}{Un(a) médico atendió el caso}
\item{Enfermero(a)_partera}{Una enfermera o partera atendió el caso}
\item{Otra}{Otro personal antedió el caso}
\item{No_especificado}{No se especifica quién atendió}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
registro_agente
}
\description{
República Mexicana. Distribución porcentual de nacimientos en niñas y adolescentes de 10 a 14 años de edad por agente que atendió el parto, 1990-2019
}
\keyword{datasets}
