% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasa_fecundidad_residencia}
\alias{tasa_fecundidad_residencia}
\title{Titulo del dataset.}
\format{
Dimensiones del dataset

\describe{
\item{Nombre Columna 1}{definition}
\item{Nombre Columna 2}{definition}
}
}
\source{
\url{https://datos.gob.mx/busca/dataset/salud-sexual-y-reproductiva/resource/9b319164-181f-4da2-a935-604c9be4b156}
}
\usage{
tasa_fecundidad_residencia
}
\description{
Descripción detallada del dataset
}
\keyword{datasets}
